requirements.txt
requests
pandas
matplotlib
streamlit
python-dateutil
statsmodels
langchain
openai